pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securedev-pentest"
version = "0.1.0"
description = "LusitAI — autonomous AI-powered penetration testing engine"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "docker>=7.0",
    "duckduckgo-search>=6.0",
    "tavily-python>=0.7.20",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.3",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "alembic>=1.14",
    "structlog>=24.0",
    "httpx>=0.27",
    "neo4j>=5.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "ruff>=0.8",
    "mypy>=1.13",
    "testcontainers>=4.0",
    "polyfactory>=2.0",
    "respx>=0.22",
    "freezegun>=1.4",
    "faker>=20.0",
    "types-PyYAML",
    "pytest-timeout>=0.14",
    "pip-audit>=2.7",
    "pre-commit>=3.0",
    "playwright>=1.40",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["pentest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: requires PostgreSQL and/or Docker daemon (deselect with '-m not integration')",
    "agent: agent tests with mocked LLM (deselect with '-m not agent')",
    "e2e: end-to-end tests requiring real services (deselect with '-m not e2e')",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
